requirements/base.txt
# requirements/base.txt - 基础依赖包
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
jinja2==3.1.2
loguru==0.7.2
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
email-validator==2.1.0

# AI相关
openai==1.3.7
anthropic==0.7.8

# 缓存和任务队列
celery
redis


# 工具库
pytz==2023.3
requests==2.31.0
pydantic[email]==2.5.0